+++ lib/Traits/TraitsClasses/TableTraits.ts
@@ -3,7 +3,6 @@ import objectTrait from "../Decorators/objectTrait";
 import primitiveArrayTrait from "../Decorators/primitiveArrayTrait";
 import primitiveTrait from "../Decorators/primitiveTrait";
 import mixTraits from "../mixTraits";
-import CatalogMemberTraits from "./CatalogMemberTraits";
 import ChartPointOnMapTraits from "./ChartPointOnMapTraits";
 import DiscretelyTimeVaryingTraits from "./DiscretelyTimeVaryingTraits";
 import ExportableTraits from "./ExportableTraits";
@@ -82,6 +81,13 @@ export default class TableTraits extends mixTraits(
     type: "boolean"
   })
   showDisableStyleOption?: boolean | undefined;
+
+  @primitiveTrait({
+    name: "Show disable time",
+    description: "If enabled, there will be an checkbox to disable time.",
+    type: "boolean"
+  })
+  showDisableTimeOption?: boolean | undefined;
 
   @primitiveArrayTrait({
     name: "Column titles",