+++ CHANGES.md
@@ -35,7 +35,6 @@ Change Log
 * `TableMixin.activeStyle` will now try to pick columns which aren't labelled `"id"` or `"_id_"` by default
 * Table columns named `"easting"` and `"northing"` are now hidden by default from styles
 * Exporting `TableMixin` will now add proper file extensions
-* Added `timeDisableDimension` to `TableMixin` - this will render a checkbox to disable time dimension if `rowGroups` only have a single time interval per group (i.e. features aren't "moving" across time)
 * Added `TimeVaryingTraits.timeLabel` trait to change label on `DateTimeSelectorSection` (defaults to "Time:")
   * This is set to `timeColumn.title`
 * `TableColumn` will try to generate prettier `title` by un-camel casing, removing underscores and capitalising words
@@ -43,6 +42,7 @@ Change Log
 * `TableColumn.type = enum` requires at least 2 unique values (including null) to be selected by default
 * `TableStyle` `startDates`, `finishDates` and `timeIntervals` will only set values for valid `rowGroups` (invalid rows will be set to `null`). For example, this means that rows with invalid regions will be ignored.
 * Add "Disable style" option to `TableMixin.styleDimensions` - it can be enabled with `TableTraits.showDisableStyleOption`
+* Added `timeDisableDimension` to `TableMixin` - this will render a checkbox to disable time dimension if `rowGroups` only have a single time interval per group (i.e. features aren't "moving" across time) - it can be enabled with `TableTraits.showDisableTimeOption`
 * Added "marker-stroke-width", "polyline-stroke-width", "polygon-stroke-width" to `GeojsonStyleTraits` (Note these are not apart of [simplestyle-spec](https://github.com/mapbox/simplestyle-spec/tree/master/1.1.0) and can only be used with `geojson-vt`)
 * Remove border from region mapping if no data
 * Tweak automatic `TableColumn.type = Enum` for wider range of values